Fix insta-off when switching back to Kodi input
When switching the input manually back to Kodi (instead of using the
service itself), it might still be asleep with the screensaver on.
In this situation, we don't want the TV to be turned off. To prevent
this we reset the screensaver activated value so that it only is
counted while the input is Kodi's.